v0.69 PR5 (partial Phase 4) — split video-utils + video-providers
Begins Plan G Phase 4 (ai-script-pipeline.ts split). 2 helper modules
extracted into commands/_shared/:

- video-utils.ts (280 L) — StoryboardSegment, retry constants, sleep,
  uploadToImgbb, extendVideoToTarget, logSceneFailure,
  waitForVideoWithRetry
- video-providers.ts (219 L) — generateVideoWithRetry{Grok,Kling,
  Runway,Veo} (4 retry helpers, depend on video-utils)

ai-script-pipeline.ts: 1,976 → 1,567 lines (−409). Re-exports kept for
backward compat (consumers in generate.ts, ai-video.ts, etc. unchanged).

## Scope of this PR

Helpers only. The two large execute functions remain in
ai-script-pipeline.ts:
- executeScriptToVideo (~800 L)
- executeRegenerateScene (~140 L)

Splitting these is a larger refactor (each is internally a multi-stage
pipeline with significant internal coupling) — future PR if motivated.

## edit_fill_gaps manifest entry — deferred

The other Phase 4 deliverable (edit_fill_gaps in manifest) requires
extracting the ~400-line .action() body in ai-fill-gaps.ts into an
`executeFillGaps` function — non-trivial because the body is tightly
coupled to ora spinners, console.log, and exitWithError CLI semantics.
Better as its own focused PR.

## Test plan

- [x] `pnpm -r exec tsc --noEmit` — 0 errors
- [x] `pnpm lint` — 0 errors
- [x] `pnpm -F @vibeframe/cli test` — 652 passed (unchanged)
- [x] All re-exports preserve external import paths
